Use generator, some refactoring
