+++ tests/FinderTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace Yiisoft\Classifier\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Classifier\ClassFinder;
+use Yiisoft\Classifier\Classifier;
 use Yiisoft\Classifier\Tests\Support\Interfaces\PostInterface;
 use Yiisoft\Classifier\Tests\Support\Interfaces\UserInterface;
 use Yiisoft\Classifier\Tests\Support\Post;
@@ -19,17 +19,21 @@ final class FinderTest extends TestCase
      */
     public function testInterfaces(string|array $interfaces, array $expectedClasses)
     {
-        $finder = new ClassFinder(__DIR__);
+        $finder = new Classifier(__DIR__);
         $finder = $finder->implements($interfaces);
 
-        $array = $finder->find();
+        $iterable = $finder->find();
 
-        $this->assertEquals($expectedClasses, $array);
+        $this->assertEquals($expectedClasses, iterator_to_array($iterable));
     }
 
     public function interfacesDataProvider(): array
     {
         return [
+            [
+                [],
+                [],
+            ],
             [
                 PostInterface::class,
                 [Post::class, PostUser::class],
